// api/src/modules/tenant/core/authorization/authorization.types.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  STAFF = 'STAFF',
  USER = 'USER',
}

export enum PermissionCode {
  /** Create/update/delete catalog products (external docs may call this PRODUCT_MANAGER). */
  PRODUCT_MANAGE = 'PRODUCT_MANAGE',
  ORDER_CREATE = 'ORDER_CREATE',
  ORDER_MANAGE = 'ORDER_MANAGE',
  PAYMENT_CREATE = 'PAYMENT_CREATE',
  PAYMENT_READ = 'PAYMENT_READ',
  PAYMENT_MANAGE = 'PAYMENT_MANAGE',
  USER_READ = 'USER_READ',
  CATEGORY_READ = 'CATEGORY_READ',
  CATEGORY_MANAGE = 'CATEGORY_MANAGE',
  INVENTORY_READ = 'INVENTORY_READ',
  INVENTORY_MANAGE = 'INVENTORY_MANAGE',
  CART_READ = 'CART_READ',
  CART_MANAGE = 'CART_MANAGE',
  ADDRESS_CREATE = 'ADDRESS_CREATE',
  ADDRESS_READ = 'ADDRESS_READ',
  ADDRESS_UPDATE = 'ADDRESS_UPDATE',
  ADDRESS_DELETE = 'ADDRESS_DELETE',
  FULFILLMENT_READ = 'FULFILLMENT_READ',
  FULFILLMENT_UPDATE = 'FULFILLMENT_UPDATE',
  RETURN_CREATE = 'RETURN_CREATE',
  RETURN_READ = 'RETURN_READ',
  RETURN_UPDATE = 'RETURN_UPDATE',
  REFUND_CREATE = 'REFUND_CREATE',
  REFUND_READ = 'REFUND_READ',
  REFUND_UPDATE = 'REFUND_UPDATE',
  NOTIFICATION_READ = 'NOTIFICATION_READ',
  NOTIFICATION_MANAGE = 'NOTIFICATION_MANAGE',
  CMS_READ = 'CMS_READ',
  CMS_EDIT = 'CMS_EDIT',
  CMS_PUBLISH = 'CMS_PUBLISH',
  PROMOTION_READ = 'PROMOTION_READ',
  PROMOTION_MANAGE = 'PROMOTION_MANAGE',
  REVIEW_READ = 'REVIEW_READ',
  REVIEW_MODERATE = 'REVIEW_MODERATE',
  WISHLIST_MANAGE = 'WISHLIST_MANAGE',
  AUDIT_READ = 'AUDIT_READ',
}

export enum PlatformPermissionCode {
  PLATFORM_TENANT_READ = 'PLATFORM_TENANT_READ',
  PLATFORM_TENANT_MANAGE = 'PLATFORM_TENANT_MANAGE',
  PLATFORM_OPERATOR_MANAGE = 'PLATFORM_OPERATOR_MANAGE',
}

export const AUTHZ_ROLES_KEY = 'authz_roles';
export const AUTHZ_PERMISSIONS_KEY = 'authz_permissions';
export const AUTHZ_OWNER_PARAM_KEY = 'authz_owner_param';